portfolio_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
WTForms==3.0.1
Werkzeug==2.3.7
email_validator